Remove WASI special handling
We used to use `std::os::wasi::ffi::OsStrExt` to treat WASI like
Unix. This has two problems:

- On the `wasm32-wasip2` the trait is unstable so this crate doesn't
  build.

- WASI doesn't actually work like Unix. Current practice is to require
  OS strings to be UTF-8. For example:

    $ wasmtime target/wasm32-wasip2/debug/examples/echo.wasm $'\xff'
    Error: failed to convert "\xFF" to utf-8

So we shouldn't have been doing this at all. We can use the existing
fallback behavior on WASI.

I started doing this because WASI is one of three targets for which
the docs list an `OsStr`, so I assumed this was important and
deliberate. But it's only there because almost all platforms re-export
Unix's `OsStrExt` and WASI is one of the three targets for which
docs are generated.